requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3